feat: Add Ability to Abort Simulation (#127)
A user can now abort a simulation or sweeping simulation by using the "abort" command under the simulation menu. Pybert will check if the thread has had its stop event set after each major simulation step.

* refactor: pull threads into their own module
* fix: add known first and third party to isort
* refactor: pull the handler into its own module
